Rename package to domain_event_broker
Use a better name for this package that's not taken yet on PyPI and
harmonize package and module name.
