+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include versioneer.py
-include domain_events/_version.py
+include domain_event_broker/_version.py
 include README.md
