+++ Makefile
@@ -2,4 +2,4 @@ test:
 	py.test -v -s
 
 coverage:
-	coverage run --source domain_events -m py.test && coverage report -m --omit=domain_events/_version.py
+	coverage run --source domain_event_broker -m py.test && coverage report -m --omit=domain_event_broker/_version.py
