+++ domain_event_broker/_version.py
@@ -38,7 +38,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = "None"
-    cfg.versionfile_source = "domain_events/_version.py"
+    cfg.versionfile_source = "domain_event_broker/_version.py"
     cfg.verbose = False
     return cfg
