+++ domain_event_broker/django/__init__.py
@@ -1,0 +1,3 @@
+from .transaction import publish_on_commit
+
+default_app_config = 'domain_event_broker.django.apps.DomainEventsConfig'
