+++ domain_event_broker/django/apps.py
@@ -1,12 +1,12 @@
 from django.apps import AppConfig
 from django.conf import settings as djsettings
 
-from domain_events import settings
+from domain_event_broker import settings
 
 
 class DomainEventsConfig(AppConfig):
-    name = 'domain_events.django'
-    label = 'domain_events_django'
+    name = 'domain_event_broker.django'
+    label = 'domain_event_broker_django'
     verbose_name = 'Domain Events'
 
     def ready(self):
