+++ domain_event_broker/tests/django/test_publish_on_commit.py
@@ -2,18 +2,18 @@ import pytest
 from mock import patch
 from django.db import transaction
 
-from domain_events.django import publish_on_commit
+from domain_event_broker.django import publish_on_commit
 
 
 @pytest.mark.django_db(transaction=True)
-@patch('domain_events.django.transaction.publish_domain_event')
+@patch('domain_event_broker.django.transaction.publish_domain_event')
 def test_publish_no_transaction(publish_mock):
     publish_on_commit('test.no_transaction', {})
     publish_mock.assert_called_with('test.no_transaction', {})
 
 
 @pytest.mark.django_db(transaction=True)
-@patch('domain_events.django.transaction.publish_domain_event')
+@patch('domain_event_broker.django.transaction.publish_domain_event')
 def test_publish_on_commit(publish_mock):
     with transaction.atomic():
         publish_on_commit('test.commit', {})
@@ -26,7 +26,7 @@ class RollbackError(Exception):
 
 
 @pytest.mark.django_db(transaction=True)
-@patch('domain_events.django.transaction.publish_domain_event')
+@patch('domain_event_broker.django.transaction.publish_domain_event')
 def test_discard_on_rollback(publish_mock):
     try:
         with transaction.atomic():
