--- domain_events/django/__init__.py
@@ -1,3 +1,0 @@
-from .transaction import publish_on_commit
-
-default_app_config = 'domain_events.django.apps.DomainEventsConfig'
