+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = domain_events.tests.django.test_settings
+DJANGO_SETTINGS_MODULE = domain_event_broker.tests.django.test_settings
