+++ setup.cfg
@@ -4,6 +4,6 @@ description-file = README.md
 [versioneer]
 VCS = git
 style = git-describe
-versionfile_source = domain_events/_version.py
-versionfile_build = domain_events/_version.py
+versionfile_source = domain_event_broker/_version.py
+versionfile_build = domain_even_broker/_version.py
 tag_prefix =
